Persist drawings server-side using Redis
Previously drawings were not persisted anywhere. Whenever a room
was refreshed or navigated to for the first time the user would not
see any of the sketches made by other users. Redis is used to
persist the drawings in an in-memory database for fast access. Once
a frame is complete the sketch needs to be transferred to AWS S3 as
a complete SVG image.